added aggregations for cppcheck

+++ src/main/java/com/fyp_poc/demo/controllers/cppCheck/CppCheckPostRequest.java
@@ -1,9 +1,6 @@
 package com.fyp_poc.demo.controllers.cppCheck;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.*;
-
-import java.util.UUID;
 
 @Getter
 @Setter
@@ -13,9 +10,9 @@ import java.util.UUID;
 public class CppCheckPostRequest {
 
     private String buildName;
-    private long error;
-    private long performance;
-    private long portability;
-    private long style;
-    private long warning;
+    private double error;
+    private double performance;
+    private double portability;
+    private double style;
+    private double warning;
 }
