+++ src/main/java/com/fyp_poc/demo/controllers/cppCheck/CppCheckResponse.java
@@ -13,9 +13,9 @@ import java.util.UUID;
 public class CppCheckResponse {
     private long id;
     private String buildName;
-    private long error;
-    private long performance;
-    private long portability;
-    private long style;
-    private long warning;
+    private double error;
+    private double performance;
+    private double portability;
+    private double style;
+    private double warning;
 }
